requirements.txt
apex==0.1
numpy==1.21.5
Pillow==8.4.0
pytorch_ignite==0.2.1
scipy==1.7.3
torch==1.8.1+cu111
torchsort==0.1.9
torchvision==0.9.1+cu111
yacs==0.1.8